Add warning for during-run import of address under control of Core.
This case can happen if you sweep to a destination under the control
of your Core wallet (although it's rare). This is not actually a problem,
since that address neither was nor will be Joinmarket's. Hence show a
warning but continue. Also add a comment to BitcoinCoreInterface.im-
port_addresses, to note that it must *only* be used for initial sync,
since it has a risk of shutting down the program.