Remove utxos from proposed transaction if spent or invalid
Previous to this patch, if an already-spent utxo was proposed
by a counterparty, the counterparty would be removed from the
proposed transaction but the input (utxo) would not, meaning
that the transaction would contain inputs for which signatures
were never going to be provided; these transactions simply time
out. This was not, of course, intended behaviour, it was intended
that the transaction would complete with the valid utxos from the
other counterparties. That's what this patch does.