(doc) Update tx_broadcast explanation
Currently, only 'self' works until the other options for tx_broadcast are implemented. Update the docs accordingly, because there was someone wondering on IRC today why his transaction failed to broadcast; the reason was he changed this setting, which currently only works for 'self'.

If the other options are implemented (tagging #368 ), this documentation should be changed along with it.