Make yg algorithms easier to define.
This refactors YieldGeneratorBasic a bit.  In particular, we move things
that custom algorithms will most likely want to change to separate
functions.  That way, it is easy to define the behaviour of inputs
and outputs for the join without the need to duplicate the other, more
general logic of create_my_orders and oid_to_order.