Add loop in the CLI for freezing UTXOs
Prior to this commit the wallet-tool (un)freeze tool would only
allow a single freeze or unfreeze per run. So a user intending
to (un)freeze several UTXOs would have to run the script several
times which would be slow due to the need to sync the wallet
many times.

See also the discussion in: http://gnusha.org/joinmarket/2020-01-09.log