Pin pytest to 5.3.5, and bitcoind download for travis
As of 13 March the latest pytest breaks the twisted tests.
As a result the version is pinned until that can be fixed.
Additionally, at the same time, the bitcoin ppa has been
removed and so cannot be used in our travis testing setup.
Hence we install bitcoin directly (for tests only, of course)
from bitcoin.org,
copying bitcoind and bitcoin-cli to `/usr/local/lib`.