Adds Windows one-click exe build via Wine
The script build/windows-build.sh can be run on Linux with
Wine pre-installed to build a self-contained .exe for JoinmarketQt.
This feature is intended for use by developers; we can add these
.exe files to future releases.
The built executable has no external dependencies and includes
libsecp256k1 (both for coincurve and for python-bitcointx) and
libsodium within it.