PySide2 version pinned to 5.14.2
Fixes #737. Before this commit, startup of
Joinmarket-Qt does not necessarily work "out
of the box" due to changes in PyQt 5.15 that
create external dependencies on shared libraries,
see comments on the Issue for details.
While we had formerly pinned PyQt 5.14, it is
in fact necessary to also pin the version of
PySide2 to the same 5.14.2 to remove the
problem.