Fix unused key-type param for importprivkey
This enables the previously ignored `key-type` parameter on the
`imprortprivkey` wallet command. Prior to this change, `key-type` was
simply ignored and the wallet would attempt to import the private key
using whatever key type the wallet was using (either native segwit or
wrapped segwit-p2sh). Now, the user can specify the format for the
imported private key, and an address/key can be imported into a wallet
using a different key type.